+++ .gitignore
@@ -1,4 +1,5 @@
 env/
 temp/
 *.csv*
+*.mp3
 *.mp4
